Made unconditional branches always taken
This is something ChampSim does.
I did not do it at the beginning because I had not noticed it
and assumed that unconditional branches would always be marked taken.
This makes one wonder whether ChampSim's method for determining
which branches are conditional is correct.
But, to have the same results as in ChampSim,
we incorporate this change.